1-backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlmodel>=0.0.14

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Email
boto3>=1.34.0
botocore>=1.34.0

# Environment
python-dotenv>=1.0.0

# HTTP & Requests
httpx>=0.25.2
requests>=2.31.0

# Data validation
pydantic>=2.5.0
email-validator>=2.1.0

# CORS
starlette>=0.27.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development/Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2  # For testing